fix: do not allow users to pass straight environ to ECS
While ECS allows the direct specification of an environ variable in the
task-definition, supporting this has allowed developers to store
passwords, tunable params, and other random data in workflows and GitHub
Secrets which are subsequently exposed "in the clear" in the task
definition. Moving forward we will require all tunable params and
secrets to be stored in SSM, Secrets Manager, or KMS and tell ECS to set
the environ via usage of `secretFrom`.